requirements.txt
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
